Added texture and normal to bottom of cone

+++ Generator/src/ConeGenerator.h
@@ -6,7 +6,7 @@
 class ConeGenerator : public AbstractGenerator {
 private:
     int bottomRadius;
-    int height;
+    float height;
     int slices;
     int stacks;
